fix(crowdstrike-remediation): minor fixes and updates to the remediation script (#25)
* fix: fix logic around mounting + update readme

Add ability to fix fs before mounting, add logging, and update readme to
account for missed IAM role

* chore: clean up logging and update readme

+++ CrowdStrike-remediate/main.py
@@ -23,6 +23,7 @@ logger = logging.getLogger("main")
 
 # Function to stop an instance
 def stop_instance(ec2_client, instance_id, dry_run=False):
+    logger.info(f"Attempting to stop instance {instance_id}")
     try:
         ec2_client.stop_instances(InstanceIds=[instance_id], DryRun=dry_run)
         waiter = ec2_client.get_waiter("instance_stopped")
@@ -34,6 +35,7 @@ def stop_instance(ec2_client, instance_id, dry_run=False):
 
 # Function to start an instance
 def start_instance(ec2_client, instance_id, dry_run=False):
+    logger.info(f"Attempting to start instance {instance_id}")
     try:
         ec2_client.start_instances(InstanceIds=[instance_id], DryRun=dry_run)
         waiter = ec2_client.get_waiter("instance_running")
@@ -45,6 +47,7 @@ def start_instance(ec2_client, instance_id, dry_run=False):
 
 # Function to detach volumes
 def detach_volumes(ec2_client, instance_id, dry_run=False):
+    logger.info(f"Attempting to detach volumes from instance {instance_id}")
     try:
         volumes = ec2_client.describe_volumes(
             Filters=[{"Name": "attachment.instance-id", "Values": [instance_id]}]
@@ -54,9 +57,11 @@ def detach_volumes(ec2_client, instance_id, dry_run=False):
             logger.error(f"Failed to get volumes for instance {instance_id}: {e}")
         return
 
+    logger.info(f"Found {len(volumes['Volumes'])} volumes to detach")
     for volume in volumes["Volumes"]:
         volume_id = volume["VolumeId"]
         try:
+            logger.info(f"Detaching volume {volume_id}")
             res = ec2_client.detach_volume(VolumeId=volume_id, DryRun=dry_run)
             waiter = ec2_client.get_waiter("volume_available")
             waiter.wait(VolumeIds=[volume_id])
@@ -69,6 +74,7 @@ def detach_volumes(ec2_client, instance_id, dry_run=False):
 
 # Function to attach volume to the current instance
 def attach_volume(ec2_client, instance_id, volume_id, device, dry_run=False):
+    logger.info(f"Attempting to attach volume {volume_id} to instance {instance_id}")
     try:
         ec2_client.attach_volume(
             InstanceId=instance_id, VolumeId=volume_id, Device=device, DryRun=dry_run
@@ -82,6 +88,7 @@ def attach_volume(ec2_client, instance_id, volume_id, device, dry_run=False):
 
 # Function to detach volume from the current instance
 def detach_volume(ec2_client, volume_id, dry_run=False):
+    logger.info(f"Attempting to detach volume {volume_id} from the current instance")
     try:
         ec2_client.detach_volume(VolumeId=volume_id, DryRun=dry_run)
         waiter = ec2_client.get_waiter("volume_available")
@@ -93,9 +100,21 @@ def detach_volume(ec2_client, volume_id, dry_run=False):
 
 # Function to remove the file
 def remove_crowdstrike_file(mount_point):
+    logger.info(f"Attempting to remove bad CrowdStrike file from {mount_point}")
     file_pattern = f"{mount_point}/Windows/System32/drivers/CrowdStrike/C-00000291*.sys"
-    for file in glob.glob(file_pattern):
-        os.remove(file)
+    files = glob.glob(file_pattern)
+    if not files:
+        logger.warning(f"No matching CrowdStrike files found in {file_pattern}")
+        return
+
+    for file in files:
+        try:
+            os.remove(file)
+            logger.info(f"Successfully removed {file}")
+        except OSError as e:
+            logger.error(f"Failed to remove {file}: {e}")
+            raise
+    logger.info("Exiting remove_crowdstrike_file")
 
 
 # Function to get the device name
@@ -104,8 +123,11 @@ def get_device_name():
     for line in res.splitlines():
         if "NTFS" in line:
             device_name = line.split(" ")[0]
-            return device_name
-    return ""
+            return device_name, "ntfs"
+        elif "HPFS/NTFS/exFAT" in line:
+            device_name = line.split(" ")[0]
+            return device_name, "ntfs"
+    return "", ""
 
 
 # Function to check we have the required pre-requisites
@@ -116,7 +138,7 @@ def check_prereq():
         raise FileNotFoundError("fdisk not found")
     if subprocess.run(["which", "ec2metadata"], capture_output=True).returncode != 0:
         raise FileNotFoundError("ec2metadata not found")
-    if subprocess.run(["ntfs-3g", "--help"], capture_output=True).returncode != 9:
+    if subprocess.run(["which", "ntfs-3g"], capture_output=True).returncode != 0:
         raise FileNotFoundError("ntfs-3g not found")
 
 
@@ -139,13 +161,13 @@ def main() -> None:
 
     region = (
         subprocess.run(
-            "ec2metadata --availability-zone", capture_output=True, shell=True
+            "ec2metadata --availability-zone", capture_output=True, shell=True, check=False
         )
         .stdout.decode()
         .strip()
     )[:-1]
     self_instance_id = (
-        subprocess.run("ec2metadata --instance-id", capture_output=True, shell=True)
+        subprocess.run("ec2metadata --instance-id", capture_output=True, shell=True, check=False)
         .stdout.decode()
         .strip()
     )
@@ -186,18 +208,29 @@ def main() -> None:
 
             try:
                 device_name = ""
+                fs_type = ""
                 for retry in range(12):
-                    device_name = get_device_name()
+                    logger.info(f"Attempt {retry + 1} to get device name")
+                    device_name, fs_type = get_device_name()
                     if device_name:
                         break
                     time.sleep(2)
-                if not dry_run:
-                    subprocess.run(
-                        ["mount", device_name, mount_point], capture_output=True
-                    )
-                    remove_crowdstrike_file(mount_point)
-                    subprocess.run(["umount", mount_point], capture_output=True)
 
+                logger.info(f"Found device name: {device_name}, FS type: {fs_type}")
+
+                if not dry_run:
+                    if fs_type == "ntfs":
+                        # Prior to mounting, let's ensure we have a clean FS
+                        logger.info("Attempting to fix NTFS filesystem prior to mounting")
+                        subprocess.run(["ntfsfix", device_name], check=False)
+                        mount_cmd = ["ntfs-3g", device_name, mount_point, "-o", "rw,force"]
+                    else:
+                        mount_cmd = ["mount", device_name, mount_point]
+                    logger.info(f"Running mount command: {mount_cmd}")
+                    subprocess.run(mount_cmd, capture_output=True, check=False)
+                    logger.info("Attempting to remove CrowdStrike bad file")
+                    remove_crowdstrike_file(mount_point)
+                    subprocess.run(["umount", mount_point], capture_output=True, check=False)
             finally:
                 detach_volume(ec2_client, volume_id, dry_run=dry_run)
                 attach_volume(
